Extract constants and types
